Undo renaming account.rs as Git chokes on the diff
